# openapi.yaml
openapi: 3.1.0
info:
  title: CodeTogether API
  version: 0.2.0
  description: |
    REST API платформы совместного редактирования кода CodeTogether.
    Все идентификаторы и поля совпадают с колонками схемы PostgreSQL из `sql/create_db.sql`.
    \nДокумент использует CRDT-журнал, который хранится в таблице `document_operation` (колонки `node_site`,
    `node_counter`, `left_node`, `right_node`, `operation_type`, `value`, `color`, `version`, `user_id`).
    Клиент обязан передавать `baseVersion` и набор операций в последовательности. Сервер применяет операции атомарно,
    увеличивает версию документа и возвращает фактический журнал.
servers:
  - url: https://api.codetogether.local/v1
    description: Прод
  - url: http://localhost:8080/v1
    description: Dev
security:
  - bearerAuth: []
...
paths:
...
  /sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
    get:
      tags: [Sessions]
      summary: Получить доску
      responses:
        '200':
          description: Информация о доске
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
    patch:
      tags: [Sessions]
      summary: Обновить имя/язык
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdateRequest'
      responses:
        '200':
          description: Обновлённые данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
    delete:
      tags: [Sessions]
      summary: Удалить доску
      responses:
        '204':
          description: Удалена
  /sessions/{sessionId}/members:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
    get:
      tags: [Sessions]
      summary: Список участников доски
      responses:
        '200':
          description: Массив участникoв
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionMember'
    post:
      tags: [Sessions]
      summary: Добавить участника (email или id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInviteRequest'
      responses:
        '201':
          description: Участник создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionMember'
  /sessions/{sessionId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
      - $ref: '#/components/parameters/UserIdParam'
    patch:
      tags: [Sessions]
      summary: Изменить роль участника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRoleRequest'
      responses:
        '200':
          description: Обновлённая роль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionMember'
    delete:
      tags: [Sessions]
      summary: Удалить участника
      responses:
        '204':
          description: Удалён
  /sessions/{sessionId}/invites:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
    post:
      tags: [Sessions]
      summary: Сгенерировать публичную ссылку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteCreateRequest'
      responses:
        '201':
          description: Возвращает токен и URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteLink'
  /invites/{token}/accept:
    parameters:
      - $ref: '#/components/parameters/TokenParam'
    post:
      tags: [Sessions]
      summary: Присоединиться по ссылке
      responses:
        '200':
          description: Доступ к доске предоставлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '410':
          description: Ссылка протухла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /public/sessions/{token}:
    parameters:
      - $ref: '#/components/parameters/TokenParam'
    get:
      tags: [Sessions]
      summary: Публичный просмотр доски (анонимный)
      security: []
      responses:
        '200':
          description: Публичные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionView'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    TokenParam:
      name: token
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
...
    SessionUpdateRequest:
      type: object
      properties:
        name:
          type: string
        language:
          type: string
    SessionSummary:
      type: object
      required: [id, name, language, ownerId, role, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        language:
          type: string
        ownerId:
          type: integer
          format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
        updatedAt:
          type: string
          format: date-time
    SessionDetails:
      allOf:
        - $ref: '#/components/schemas/SessionSummary'
        - type: object
          required: [link, linkExpiresAt]
          properties:
            link:
              type: string
            linkExpiresAt:
              type: string
              format: date-time
            stats:
              $ref: '#/components/schemas/DocumentStats'
...
    SessionMember:
      type: object
      required: [sessionId, userId, role, joinedAt]
      properties:
        sessionId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
        joinedAt:
          type: string
          format: date-time
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
    MemberInviteRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        userId:
          type: integer
          format: int64
        role:
          type: string
          enum: [editor, viewer]
          default: editor
      anyOf:
        - required: [email]
        - required: [userId]
    MemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, editor, viewer]
    InviteCreateRequest:
      type: object
      required: [expiresInMinutes]
      properties:
        expiresInMinutes:
          type: integer
          minimum: 5
          maximum: 1440
    InviteLink:
      type: object
      required: [sessionId, link, expiresAt]
      properties:
        sessionId:
          type: integer
          format: int64
        link:
          type: string
        expiresAt:
          type: string
          format: date-time
    PublicSessionView:
      type: object
      required: [sessionId, name, language, ownerName, document]
      properties:
        sessionId:
          type: integer
          format: int64
        name:
          type: string
        language:
          type: string
        ownerName:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        document:
          $ref: '#/components/schemas/DocumentState'
    DocumentState:
      type: object
      required: [id, sessionId, version, content]
      properties:
        id:
          type: integer
          format: int64
        sessionId:
          type: integer
          format: int64
        version:
          type: integer
        content:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    DocumentStats:
      type: object
      required: [sessionId, documentId, activeParticipants, operationCount, lastSnapshotVersion]
      properties:
        sessionId:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        activeParticipants:
          type: integer
        operationCount:
          type: integer
        lastSnapshotVersion:
          type: integer
        averageLatencyMs:
          type: integer
          nullable: true
        typingRate:
          type: number
          format: float
          nullable: true